# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /demo/prototype:
    post:
      operationId: generateDemoPrototype
      tags: [demo]
      summary: Generate a quick AI prototype from a short prompt
      description: Generates a single-page HTML prototype preview from a visitor prompt (max 200 chars) and optional reference images. Rate limited per IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoPromptInput"
      responses:
        "201":
          description: Prototype generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoPrototypeResult"
        "400":
          description: Invalid prompt
        "429":
          description: Too many requests
  /demo/submissions:
    post:
      operationId: submitDemoInquiry
      tags: [demo]
      summary: Send a liked prototype to the team with contact details
      description: Stores the demo request as a project inquiry linked to the generated prototype and notifies the team.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoSubmissionInput"
      responses:
        "201":
          description: Submission received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InquiryStatus"
        "400":
          description: Invalid submission
        "429":
          description: Too many requests
...
components:
  schemas:
...
    DemoPromptInput:
      type: object
      properties:
        prompt: { type: string, minLength: 5, maxLength: 200 }
        images:
          type: array
          maxItems: 2
          description: Optional reference images (logo, sketch, inspiration) the prototype should be based on.
          items:
            type: object
            properties:
              mediaType:
                type: string
                enum: [image/jpeg, image/png, image/webp, image/gif]
              data:
                type: string
                maxLength: 2800000
                description: Base64-encoded image bytes (no data URL prefix). Max ~2MB per image.
            required:
              - mediaType
              - data
        website:
          type: string
          description: Honeypot field — must be empty; bots that fill it are silently dropped.
      required:
        - prompt
    DemoPrototypeResult:
      type: object
      properties:
        prototypeId: { type: integer }
        html: { type: string }
      required:
        - prototypeId
        - html
    DemoSubmissionInput:
      type: object
      properties:
        prototypeId: { type: integer }
        prompt: { type: string, minLength: 5, maxLength: 200 }
        name: { type: string, minLength: 1 }
        email: { type: string, minLength: 3 }
        company: { type: string }
        phone: { type: string, minLength: 5 }
        budget: { type: string }
        timeline: { type: string }
        details: { type: string }
        website:
          type: string
          description: Honeypot field — must be empty; bots that fill it are silently dropped.
      required:
        - prototypeId
        - prompt
        - name
        - email
        - phone
...
    InquiryStatus:
      type: object
      properties:
        status:
          type: string
          enum: [received]
        referenceNumber:
          type: string
          description: Public lead reference number shown to the visitor and used in emails.
      required:
        - status
...